apps/ActionOS/requirements-lambda.txt
# Lambda deployment dependencies
# boto3 excluded — provided by Lambda runtime
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
requests>=2.28.0
pywebpush>=2.0.0
